+++ tests/test_cid_disjoint.cpp
@@ -147,8 +147,8 @@ void (* test_impls []) (Madara::Cid::Settings &) = {
 
 };
 
-unsigned long long clocks [NUM_TESTS + 1];
-unsigned long long latencies [NUM_TESTS];
+uint64_t clocks [NUM_TESTS + 1];
+uint64_t latencies [NUM_TESTS];
 
 void print_averages (std::ostream & output, 
                      Madara::Cid::Settings & settings)
@@ -552,7 +552,7 @@ void verify_algorithms (std::ostream & output)
   #define DEGREED_GA              4
   #define NAIVE_GA                5
 
-  unsigned long long alg_latencies[CHECKED_ALGORITHMS] = {0,0,0,0,0,0};
+  uint64_t alg_latencies[CHECKED_ALGORITHMS] = {0,0,0,0,0,0};
 
   std::string divider (79, '*');
   output << divider << std::endl;
@@ -648,7 +648,7 @@ void verify_algorithms (std::ostream & output)
 
   Madara::Cid::generate_worst_solution (settings);
 
-//  unsigned long long actual_latency, ideal_latency = 0, worst_latency;
+//  uint64_t actual_latency, ideal_latency = 0, worst_latency;
 
   alg_latencies[WORST_LATENCY] = Madara::Cid::calculate_latency (settings);
